+++ lib/src/tools/expando_ext.dart
@@ -1,0 +1,11 @@
+extension ExpandoGetOrPutExt<T extends Object> on Expando<T> {
+  T getOrPut(Object? key, T Function() defaultValue) {
+    if (key == null) return defaultValue();
+    T? result = this[key];
+    if (result == null) {
+      result = defaultValue();
+      this[key] = result;
+    }
+    return result;
+  }
+}